// src/core/models/role.model.ts
// Generated by https://quicktype.io

export interface Role {
  rolId: string
  name: string
  permissions: string[]
  createdAt: string
  updatedAt: string
}

export const RoleNames: { [key: string]: string } = {
  MANAGER: 'Manager',
  USER: 'User',
  SUPER_ADMIN: 'Super Admin',
  SYSTEM_ADMIN: 'System Admin',
  CONTRIBUTOR: 'Contributor'
}

export enum RoleEnum {
  MANAGER = 'MANAGER',
  USER = 'USER',
  SUPER_ADMIN = 'SUPER_ADMIN',
  SYSTEM_ADMIN = 'SYSTEM_ADMIN',
  CONTRIBUTOR = 'CONTRIBUTOR'
}